[move]: move firebase_auth_service
